+++ src/client/actions/item.js
@@ -2,12 +2,56 @@ import {
   API,
   CHANGE_ITEM,
   DELETE_ITEM,
-  FETCH_ITEM,
   FETCH_UPLOAD,
-  // MAYBE_SAVE_ITEM,
-  RESET_ITEM
-  // UPDATE_ITEM
+  UPDATE_ITEM
 } from '../actions'
+
+import fetch from 'isomorphic-fetch'
+import * as url from 'url'
+const { API_URL } = process.env
+
+export const FETCH_ITEM_SUCCESS = 'FETCH_ITEM_SUCCESS'
+export const FETCH_ITEM_ERROR = 'FETCH_ITEM_ERROR'
+export const FETCH_ITEM_REQUESTED = 'FETCH_ITEM_REQUESTED'
+export const RESET_ITEM = 'RESET_ITEM'
+
+export const fetchItem = (model = '', id, query = {}) => dispatch => {
+  const encodedURI = url.parse(`${API_URL}${model}/${id}`)
+  encodedURI.query = query
+  const formattedURI = url.format(encodedURI)
+
+  dispatch({
+    type: FETCH_ITEM_REQUESTED
+  })
+
+  return fetch(formattedURI)
+    .then(res => {
+      if (res) {
+        if (!res.ok) {
+          throw Error(res.status)
+        }
+        return res.json()
+      }
+    })
+    .then(item => {
+      dispatch({
+        type: FETCH_ITEM_SUCCESS,
+        payload: {
+          item
+        }
+      })
+      return item
+    })
+    .catch(error => {
+      dispatch({
+        type: FETCH_ITEM_ERROR,
+        payload: {
+          error: { message: error.toString() }
+        }
+      })
+      return null
+    })
+}
 
 export const changeItem = (key, value) => {
   return {
@@ -19,22 +63,9 @@ export const changeItem = (key, value) => {
   }
 }
 
-export const fetchItem = (model, id) => {
-  return {
-    type: API,
-    payload: {
-      id,
-      method: 'get',
-      model,
-      next: FETCH_ITEM,
-      url: `/${model}/${id}`
-    }
-  }
-}
-
 export const maybeSaveItem = (model, forceSave) => {
   return (dispatch, getState) => {
-    const { item } = getState().item
+    const { item } = getState().itemReducer
 
     if (!item.published || forceSave) {
       dispatch(updateItem(model, item))
