--- src/client/apps/index.js
@@ -1,11 +1,0 @@
-export { default as Event } from './event'
-export { default as Events } from './events'
-export { default as Pages } from './pages'
-export { default as Project } from './project'
-export { default as Projects } from './projects'
-export { default as Publication } from './publication'
-export { default as Publications } from './publications'
-export { default as Settings } from './settings'
-export { default as NewUser } from './user/new'
-export { default as Login } from './user/login'
-export { default as Logout } from './user/logout'
