+++ src/client/routes.js
@@ -1,67 +1,99 @@
-import styled, { ThemeProvider } from 'styled-components'
-import React, { Component } from 'react'
-import { Route } from 'react-router-dom'
-import { theme } from './styles/theme.jsx'
-import {
-  Event,
-  Events,
-  Login,
-  Logout,
-  Pages,
-  Project,
-  Projects,
-  Publication,
-  Publications,
-  Settings,
-  NewUser
-} from './apps'
-import Header from './components/header/header.jsx'
+import { fetchItem } from 'client/actions/item'
+import { fetchItems } from 'client/actions/items2'
+import { fetchPage } from 'client/actions/page'
+import Item from 'client/apps/Item/Item'
+import Items from 'client/apps/Items/Items'
+import Page from 'client/apps/pages/Page'
+import Login from 'client/apps/user/login'
 
-export default class Routes extends Component {
-  render () {
-    return (
-      <ThemeProvider theme={theme}>
-        <Content>
-          <Header />
-          <Main>
-            <Route exact path='/' component={Events} />
-            <Route exact path='/events' component={Events} />
-            <Route path='/events/:id' component={Event} />
+const { HOMEPAGE_ENABLED } = process.env
 
-            <Route exact path='/projects' component={Projects} />
-            <Route exact path='/projects/:id' component={Project} />
+const getQuery = store => {
+  if (!store.getState().user.isAuthenticated) {
+    return { published: true }
+  }
+  return {}
+}
 
-            <Route exact path='/releases' component={Publications} />
-            <Route exact path='/releases/:id' component={Publication} />
-            <Route exact path='/publications' component={Publications} />
-            <Route exact path='/publications/:id' component={Publication} />
-
-            <Route exact path='/login' component={Login} />
-            <Route exact path='/logout' component={Logout} />
-            <Route exact path='/new/user' component={NewUser} />
-            <Route exact path='/info' component={Pages} />
-            <Route exact path='/settings' component={Settings} />
-          </Main>
-        </Content>
-      </ThemeProvider>
-    )
+const HomeRoute = {
+  path: '/',
+  exact: true,
+  component: HOMEPAGE_ENABLED ? Page : Items,
+  model: HOMEPAGE_ENABLED ? undefined : 'events',
+  title: 'Home',
+  fetchInitialData: (path = '', store) => {
+    if (HOMEPAGE_ENABLED) {
+      return store.dispatch(fetchPage('/home'))
+    } else {
+      return store.dispatch(fetchItems('/events', getQuery(store)))
+    }
   }
 }
 
-const Main = styled.main`
-  padding-top: 2.5em;
-`
-
-const Content = styled.div`
-  font-family: 'Roboto', 'helvetica neue', 'helvetica', 'arial', 'sans-serif';
-  font-size: 16px;
-  line-height: 1.2em;
-  letter-spacing: .015em;
-
-  a {
-    color: black;
-    &:hover {
-      color: #ddd;
+export const routes = [
+  HomeRoute,
+  {
+    path: '/events/:id',
+    model: 'events',
+    component: Item,
+    fetchInitialData: (path = '', store) => {
+      return store.dispatch(fetchItem('/events', path.split('/').pop(), getQuery(store)))
+    }
+  },
+  {
+    path: '/events',
+    component: Items,
+    model: 'events',
+    title: 'Events',
+    fetchInitialData: (path = '', store) => {
+      return store.dispatch(fetchItems(path, getQuery(store)))
+    }
+  },
+  {
+    path: '/info',
+    component: Page,
+    title: 'Info',
+    fetchInitialData: (path = '', store) => {
+      return store.dispatch(fetchPage(path))
+    }
+  },
+  {
+    path: '/login',
+    component: Login,
+    title: 'Login'
+  },
+  {
+    path: '/projects/:id',
+    model: 'projects',
+    component: Item,
+    fetchInitialData: (path = '', store) => {
+      return store.dispatch(fetchItem('/projects', path.split('/').pop(), getQuery(store)))
+    }
+  },
+  {
+    path: '/projects',
+    component: Items,
+    model: 'projects',
+    title: 'Projects',
+    fetchInitialData: (path = '', store) => {
+      return store.dispatch(fetchItems(path, getQuery(store)))
+    }
+  },
+  {
+    path: '/releases/:id',
+    model: 'publications',
+    component: Item,
+    fetchInitialData: (path = '', store) => {
+      return store.dispatch(fetchItem('/publications', path.split('/').pop(), getQuery(store)))
+    }
+  },
+  {
+    path: '/releases',
+    component: Items,
+    model: 'publications',
+    title: 'Releases',
+    fetchInitialData: (path = '', store) => {
+      return store.dispatch(fetchItems('/publications', getQuery(store)))
     }
   }
-`
+]
